Added tests for larmor fins plc
